AsistenciaModulo Terminado

+++ routes/web.php
@@ -142,5 +142,5 @@ Route::delete('/admin/pagos/{id}', [App\Http\Controllers\PagoController::class,'
 Route::get('/admin/asistencias', [App\Http\Controllers\AsistenciaController::class,'index'])->name('admin.asistencias.index')->middleware('auth', 'can:admin.asistencias.index');
 Route::get('/admin/asistencias/create/asignacion/{id}', [App\Http\Controllers\AsistenciaController::class,'create'])->name('admin.asistencias.create')->middleware('auth', 'can:admin.asistencias.create');
 Route::post('/admin/asistencias/create', [App\Http\Controllers\AsistenciaController::class,'store'])->name('admin.asistencias.store')->middleware('auth', 'can:admin.asistencias.store');
-Route::put('/admin/asistencias/{id}', [App\Http\Controllers\AsistenciaController::class,'update'])->name('admin.asistencias.update')->middleware('auth');
-Route::delete('/admin/asistencias/{id}', [App\Http\Controllers\AsistenciaController::class,'destroy'])->name('admin.asistencias.destroy')->middleware('auth');
+Route::put('/admin/asistencias/{id}', [App\Http\Controllers\AsistenciaController::class,'update'])->name('admin.asistencias.update')->middleware('auth', 'can:admin.asistencias.update');
+Route::delete('/admin/asistencias/{id}', [App\Http\Controllers\AsistenciaController::class,'destroy'])->name('admin.asistencias.destroy')->middleware('auth', 'can:admin.asistencias.destroy');
